# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pannaga API",
    "description": "Hyperlocal Monsoon Onset & Break Prediction System for India",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/geo/tiles/blocks/{z}/{x}/{y}.pbf": {
      "get": {
        "tags": [
          "Geography & Maps"
        ],
        "summary": "MVT tiles for blocks (low zoom)",
        "description": "Generate Mapbox Vector Tile (MVT) for blocks at tile coordinate (z, x, y).\nJoins block_forecast_agg for fast rendering at low zoom levels (zoom 4 - 10).",
        "operationId": "get_block_mvt_tile_api_v1_geo_tiles_blocks__z___x___y__pbf_get",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          },
          {
            "name": "lead_week",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 1,
              "default": 1,
              "title": "Lead Week"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/geo/tiles/panchayats/{z}/{x}/{y}.pbf": {
      "get": {
        "tags": [
          "Geography & Maps"
        ],
        "summary": "MVT tiles for panchayats (high zoom)",
        "description": "Generate Mapbox Vector Tile (MVT) for Gram Panchayats at tile coordinate (z, x, y).\nProvides hyperlocal probability attributes for 3D extrusion at zoom 10 - 18.",
        "operationId": "get_panchayat_mvt_tile_api_v1_geo_tiles_panchayats__z___x___y__pbf_get",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          },
          {
            "name": "lead_week",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 1,
              "default": 1,
              "title": "Lead Week"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/geo/geojson/panchayats": {
      "get": {
        "tags": [
          "Geography & Maps"
        ],
        "summary": "GeoJSON FeatureCollection for panchayats (2D Fallback)",
        "description": "Returns GeoJSON FeatureCollection of panchayats with forecast properties.\nUsed by ChoroplethMap2D fallback when WebGL is unavailable.",
        "operationId": "get_panchayats_geojson_api_v1_geo_geojson_panchayats_get",
        "parameters": [
          {
            "name": "lead_week",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 1,
              "default": 1,
              "title": "Lead Week"
            }
          },
          {
            "name": "block_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Block Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/geo/search": {
      "get": {
        "tags": [
          "Geography & Maps"
        ],
        "summary": "Search geography by query string",
        "description": "Search panchayats, blocks, and districts. Returns coordinates and hierarchy.",
        "operationId": "search_geo_api_v1_geo_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/geo/lookup": {
      "get": {
        "tags": [
          "Geography & Maps"
        ],
        "summary": "Point-in-polygon lookup from lat/lon",
        "description": "Finds which panchayat, block, and district contains (lat, lon).",
        "operationId": "point_in_polygon_lookup_api_v1_geo_lookup_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 38.0,
              "minimum": 6.0,
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 98.0,
              "minimum": 68.0,
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}